// backend/src/common/enums/service.enum.ts
export enum ServiceCategory {
  PROGRAMMING = 'programming',
  DESIGN = 'design',
  DIGITAL_MARKETING = 'digital_marketing',
  WRITING_TRANSLATION = 'writing_translation',
  VIDEO_AUDIO = 'video_audio',
  BUSINESS = 'business',
  DATA_ANALYSIS = 'data_analysis',
  AI = 'ai',
  ENGINEERING = 'engineering',
  LIFESTYLE = 'lifestyle',
}

export enum ServiceStatus {
  PENDING = 'pending',
  PUBLISHED = 'published',
  REJECTED = 'rejected',
}